[ENH] Add remove_learnware in user.py

+++ api/user.py
@@ -16,6 +16,9 @@ user_api = Blueprint("User-API", __name__)
 
 def remove_learnware(learnware_id: str) -> bool:
     # [TODO] Require code for engine
+    ok = C.engine.delete_learnware(learnware_id)
+    if not ok:
+        return False
     cnt = database.remove_learnware("learnware_id", learnware_id)
     return cnt > 0
